APACHE-KYLIN-2732: better way to fix bug of missing records

+++ core-cube/src/main/java/org/apache/kylin/cube/inmemcubing/ConsumeBlockingQueueController.java
@@ -78,7 +78,7 @@ public class ConsumeBlockingQueueController<T> implements Iterator<T> {
         hasException = true;
     }
 
-    public long getOutputRowCount() {
+    public int getOutputRowCount() {
         return outputRowCount.get();
     }
 }
